requirements.txt
aiogram~=3.20.0.post0
gspread==6.2.0
oauth2client==4.1.3
pytesseract==0.3.13
opencv-python-headless==4.11.0.86
python-dotenv==1.1.0
